Try Quick Image Src Fix
Github pages seems to not like relative paths. In my research it seems that installing a Gemfile called jekyll would allow for using different paths for local and live hosting. That's too much effort for what will ultimately be only a few images. I've changed the path, if that doesn't work I'll just link directly to the image's original sources